Split MonadSqlQuery transactions into new class
The MonadSqlQuery class is a bit overloaded, in that it represents both
monads that can execute SQL queries with `SqlQueryRep`, and also monads
that can execute transactions. These classes do not have to coincide,
and in fact there are many monads which are likely to be one but not the
other (speaking from experience.)

Without re-arranging any of the existing instances, this simply splits
up the `MonadSqlQuery` class into two smaller classes managing each of
these responsibilities, and does the same for each instance. The code
should be functionally unchanged.

In the future, I will want to reconsider which monads fall into each
bucket and how to handle `TransactionM`, which is aggressive about
lifting when we may in fact want certain effects in the monad stack.