Make shimmed functions run on SqlBackend-compatible backends
This adds in the notion of backend compatibility, based on Persistent's
similar `BackendCompatible`. Compatibility is expressed in terms of
query representations, and allows a single backend to run queries of
multiple different types using its compatibility instances.

This is useful to us because it allows for the `Shim` functions to run
against backends that are not necessarily *exactly* `SqlBackend`. In the
MTL style, we can declare a constraint on our query `Backend` (namely,
`QueryRepCompatible SqlBackend`) and have automatic compatibility with a
host of backend representations. This avoids having to manually use
`withCompatibleQueryRep` everywhere (which, in the MTL analogy, would be
like using `lift` all the time.)