Add runVia helper for backend polymorphism
This polymorphic helper is useful for cases where the current backend
type is constrained to be compatible with a particular concrete backend,
and the programmer wants to call a function which relies on
compatibility which that *concrete backend* satisfies.

This helps avoid having to constrain the backend type twice when one of
the constraints is a strict superset of the other - in particular, when
constraining by compatibility with two concrete types which are
themselves compatible.

This came up for us when switching over to `RawPostgresql`, since the
streaming functions constrained the backend to compatibility with that
wrapper but also wanted to call `MonadSqlQuery` functions that
constrained to compatibility with the inner `SqlBackend`.

Note that this newtype can't be replaced with an ambiguous typeclass
instance, both because of coherence issues and because the compiler
would have no guidance on how to direct its type search. Because
`QueryRepCompatible` is so generic, we need to add info for the compiler
to use.